Conflicts:
	voyageone-service2/voyageone-service2-impl/src/main/java/com/voyageone/service/impl/cms/sx/SxProductService.java
	voyageone-service2/voyageone-service2-impl/src/main/java/com/voyageone/service/impl/cms/sx/rule_parser/MasterWordParser.java
	voyageone-service2/voyageone-service2-impl/src/main/java/com/voyageone/service/impl/cms/sx/word/CustomWordModuleGetCommonImages.java
	voyageone-service2/voyageone-service2-impl/src/main/java/com/voyageone/service/impl/cms/sx/word/CustomWordModuleGetMainPropductImages.java
	voyageone-task2/task2-cms/src/main/java/com/voyageone/task2/cms/service/CmsBuildPlatformProductUploadJdService.java
	voyageone-task2/task2-cms/src/main/java/com/voyageone/task2/cms/service/CmsBuildPlatformProductUploadTmMqService.java

+++ voyageone-service2/voyageone-service2-model/src/main/java/com/voyageone/service/model/cms/CmsBtSxWorkloadModel.java
@@ -15,7 +15,7 @@ import com.voyageone.base.dao.mysql.BaseModel;
 public class CmsBtSxWorkloadModel extends BaseModel {
     protected String channelId;
 
-    protected Long groupId;
+    protected Integer groupId;
 
     protected Integer publishStatus;
 
@@ -29,11 +29,11 @@ public class CmsBtSxWorkloadModel extends BaseModel {
         this.channelId = channelId == null ? null : channelId.trim();
     }
 
-    public Long getGroupId() {
+    public Integer getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(Long groupId) {
+    public void setGroupId(Integer groupId) {
         this.groupId = groupId;
     }
